+++ apps/www/src/app/(main)/(docs)/components/file-upload-single/examples/avatar.tsx
@@ -1,0 +1,9 @@
+/* eslint-disable no-console */
+
+"use client";
+
+import { FileUploadUsage } from "./usage";
+
+export const FileUploadAvatar = () => {
+  return <FileUploadUsage isAvatar />;
+};